move event listener setup to handlers.ts, add disposal function
