+++ src/featureFlags/handlers.ts
@@ -8,6 +8,27 @@ import { FEATURE_FLAG_DEFAULTS, FeatureFlags } from "./constants";
 
 const logger = new Logger("featureFlags.handlers");
 
+/**
+ * Sets up the event listeners for the LaunchDarkly client.
+ * @param client The LaunchDarkly client.
+ */
+export function setEventListeners(client: LDElectronMainClient): void {
+  // client.on doesn't return a listener, so we have to rely on the extension's deactivate() to
+  // handle the client cleanup
+  client.on("ready", () => handleClientReady(client));
+
+  client.on("failed", (err) => {
+    logger.error("failed event:", err);
+  });
+
+  client.on("error", (err) => {
+    logger.error("error event:", err);
+  });
+
+  // this is the main one we care about after the client is ready:
+  client.on("change", handleFlagChanges);
+}
+
 /** Callback function for handling "ready" events from the LD stream, which sets up the initial
  * feature flag values from LaunchDarkly, overriding any default values set during activation. */
 export async function handleClientReady(client: LDElectronMainClient) {
